+++ lib/zipkin-tracer/trace.rb
@@ -8,6 +8,13 @@ module Trace
     @tracer
   end
 
+  def self.with_trace_id(trace_id, &block)
+    self.push(trace_id)
+    yield
+  ensure
+    self.pop
+  end
+
   # A span may contain many annotations
   # This class is defined in finagle-thrift. We are adding extra methods here
   class Span