+++ spec/lib/trace_client_spec.rb
@@ -107,4 +107,17 @@ describe ZipkinTracer::TraceClient do
       end
     end
   end
+
+  describe 'Local tracing spans are nesting' do
+    it 'have same parent_id but different span_ids' do
+      subject.local_component_span(lc_value) do |ztc|
+        parent_local_trace = Trace.id
+        subject.local_component_span(lc_value) do |ztc|
+          expect(parent_local_trace.trace_id).to eq(Trace.id.trace_id)
+          expect(parent_local_trace.span_id).not_to eq(Trace.id.span_id)
+          expect(Trace.id.parent_id).to eq(parent_local_trace.span_id)
+        end
+      end
+    end
+  end
 end