+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.12.1
-Nest local tracing spans
+Allow nesting of local tracing spans
 
 # 0.12.0
 Add local tracing, fix flushing, add timestamp and duration to span